Reove qtimeseries service and fix error in case of no layer are qtimeseried
